Adding alert box for delete password

+++ src/app/Components/password-manager-summary/password-manager-summary.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { PasswordManagerServiceService } from '../../Service/password-manager-service.service'
 import { AddPasswordComponent } from '../add-password/add-password.component';
 import { NgbModal, NgbModalOptions } from '@ng-bootstrap/ng-bootstrap';
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-password-manager-summary',
   templateUrl: './password-manager-summary.component.html',
@@ -29,12 +30,24 @@ export class PasswordManagerSummaryComponent implements OnInit {
   }
 
   DeletePassword(id:any) {
-    this.passwordService.Delete("DeletePassword?id=" + id).subscribe(res => {
-      if (res.statusCode == 'SUCCESS') {
-        this.GetAllPassword();
-      }
-    }
-    )
+    Swal.fire({
+      title: 'Are you sure to delete the password?',
+      text: 'This process is irreversible.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, go ahead.',
+      cancelButtonText: 'No, let me think',
+    }).then((result) => {
+      if (result.value) {
+        this.passwordService.Delete("DeletePassword?id=" + id).subscribe(res => {
+          if (res.statusCode == 'SUCCESS') {
+           Swal.fire('Removed!', res.statusText, 'success');
+            this.GetAllPassword();
+          }
+        }
+        )
+      } 
+    });
   }
 
   GetPasswordDetailById(passwordManagerId: any) {
@@ -42,6 +55,10 @@ export class PasswordManagerSummaryComponent implements OnInit {
       if (res.statusCode == 'SUCCESS') {
         this.PasswordDetails = res.data;
         this.openModal(this.PasswordDetails)
+      }
+      else
+      {
+        Swal.fire("FAILURE",res.statusText,'error')
       }
     }
     )
